% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_footer.R
\name{add_footer}
\alias{add_footer}
\title{Add a footer with the Exploristics logo to a ggplot2 figure.}
\usage{
add_footer(filename = NULL, text = NULL, line = F, logo = NULL)
}
\arguments{
\item{filename}{An image file of a ggplot2.}

\item{text}{Optional text to add to the left of the footer. Single line or across 2 lines if you add \verb{\\n} between the lines of text.}

\item{line}{Add a line between the plot and the footer. Defaults to \code{FALSE}.}

\item{logo}{Path to an image file to use in the footer. The path must be URL, filename or raw vector. Defaults uses the Exploristics logo supplied with the package.}
}
\description{
This function returns a ggplot with a footer with the Exploristics logo and an optional text caption added below the plot.
}
\details{
This function returns a ggplot with a footer with the Exploristics logo at the bottom-right corner and an optional text caption at the bottom-left corner below the plot.
There is also an option to add a line between the plot and the footer.
The image file can be generated by creating a plot the running \code{ggsave()} with a filename specified.
The plot with the footer added is saved to a .png file with the same filename as the original plot but with the suffix "with_footer" added to it.
}
\examples{
library(ggplot2)
library(magick)
library(magrittr)

## generate a plot with the Exploristics theme
## if you want to use a numeric variable as discrete use `as.factor()` when setting `fill`
cars_plot <- ggplot(data = mtcars, aes(x = cyl, y = mpg, fill = as.factor(cyl))) +
geom_bar(stat="identity") +
exploristics_theme()

## add colour scheme and wrap text labels
cars_plot \%>\%
exploristics_fill(colour_pal="Expl_External") \%>\%
text_wrapper()

## save the plot
ggsave(filename = "example_cars_plot.png", width = 8, height = 8, dpi = 300)

## add the footer to the saved plot
add_footer("example_cars_plot.png", text= "Source:data source\nProduced by: Name")
}
\seealso{
\code{\link[magick]{image_read}}

\code{\link[magick]{image_trim}}

\code{\link[magick]{image_scale}}

\code{\link[magick]{image_background}}

\code{\link[magick]{image_extent}}

\code{\link[magick]{image_border}}

\code{\link[magick]{image_annotate}}

\code{\link[magick]{image_info}}

\code{\link[magick]{image_blank}}

\code{\link[magick]{image_append}}

\code{\link[magick]{image_write}}
}
