% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Expl_Blue}
\alias{Expl_Blue}
\title{Exploristics Blue}
\format{
An object of class \code{character} of length 6.
}
\usage{
Expl_Blue
}
\description{
A character vector contain hex colour codes for Exploristics Blues.
}
\keyword{datasets}
