% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Expl_Brand}
\alias{Expl_Brand}
\title{Colour Pallette Using Exploristics Brand Colours}
\format{
An object of class \code{character} of length 7.
}
\usage{
Expl_Brand
}
\description{
Colour Pallette Using Exploristics Brand Colours
}
\details{
Contains the following hex colours
\itemize{
\item{#18213B}
\item{#718BDE}
\item{#A4B8DC}
\item{#89D8C1}
\item{#DEF58C}
\item{#EFC3B1}
\item{#CD8581}
}
}
\keyword{datasets}
