% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Expl_External}
\alias{Expl_External}
\title{Exploristics External}
\format{
An object of class \code{character} of length 6.
}
\usage{
Expl_External
}
\description{
A character vector contain hex colour codes using External Colour Palette.
}
\keyword{datasets}
