% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Expl_HighCont}
\alias{Expl_HighCont}
\title{Exploristics High Contrast}
\format{
An object of class \code{character} of length 6.
}
\usage{
Expl_HighCont
}
\description{
A character vector contain hex colour codes for High Contrast for use
with accessibility in mind.
}
\keyword{datasets}
