% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Expl_Mono}
\alias{Expl_Mono}
\title{Colour a palette of 6 colours ranging from a dark navy through grey to white}
\format{
An object of class \code{character} of length 6.
}
\usage{
Expl_Mono
}
\description{
Colour a palette of 6 colours ranging from a dark navy through grey to white
}
\details{
Contains the following hex colours
\itemize{
\item{#121a31"}
\item{#16213d"}
\item{#2d3750"}
\item{#5c6477"}
\item{#ecf1f8"}
\item{#ffffff"}
}
}
\keyword{datasets}
