% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Expl_Rebrand}
\alias{Expl_Rebrand}
\title{Exploristics Rebrand}
\format{
An object of class \code{character} of length 4.
}
\usage{
Expl_Rebrand
}
\description{
A character vector contain hex colour codes using Rebrand Colour Palette.
}
\keyword{datasets}
