% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Expl_RGB}
\alias{Expl_RGB}
\title{Colour Palettes Using Exploristics RGB}
\format{
An object of class \code{character} of length 6.
}
\usage{
Expl_RGB
}
\description{
Colour Palettes Using Exploristics RGB
}
\details{
The hex colours for the 3 palettes
}
\keyword{datasets}
