% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Expl_RGB_dark}
\alias{Expl_RGB_dark}
\title{Colour palette of darker shades of the 6 colours within \code{Expl_RGB}}
\format{
An object of class \code{character} of length 6.
}
\usage{
Expl_RGB_dark
}
\description{
Colour palette of darker shades of the 6 colours within \code{Expl_RGB}
}
\details{
Contains the following hex colours
\itemize{
\item{#576fb8"}
\item{#ae6867"}
\item{#b1ca65"}
\item{#4c9d88"}
\item{#7f93b3"}
\item{#ca9e8e"}
}
}
\keyword{datasets}
