% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploristics_color.R
\name{exploristics_color}
\alias{exploristics_color}
\title{Add Exploristics colour to ggplot chart}
\usage{
exploristics_color()
}
\description{
This function allows you to add the Exploristics colour to your ggplotgraphics.
This function sets the colour of the graphs to a gradient matching the exploristics colour scheme
}
\keyword{exploristics_colour}
