% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploristics_style.R
\name{exploristics_style}
\alias{exploristics_style}
\title{Add Exploristics theme to ggplot chart}
\usage{
exploristics_style()
}
\description{
This function allows you to add the Exploristics theme to your ggplotgraphics.
This function takes no arguments and is added to your ggplot to set axis, background, and title look to one matching the company's style.
}
\keyword{exploristics_style}
