% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploristics_theme.R
\name{exploristics_theme}
\alias{exploristics_theme}
\title{Add Exploristics theme to ggplot2 figure.}
\usage{
exploristics_theme(
  axis_lines = TRUE,
  grid_lines = FALSE,
  legend = TRUE,
  legend_side = TRUE,
  legend_title = TRUE,
  title_size = 20,
  subtitle_size = 16,
  axis_title_size = 18,
  axis_size = 18
)
}
\arguments{
\item{axis_lines}{Add x- and y-axis lines to the plot. Defaults to \code{TRUE}.}

\item{grid_lines}{Add major grid lines to the plot. Defaults to \code{FALSE}.}

\item{legend}{Display plot legend. Defaults to \code{TRUE}.}

\item{legend_side}{Position the legend to the right side id \code{TRUE} or the bottom of the plot if \code{FALSE}. Defaults to \code{TRUE}.}

\item{legend_title}{Display legend title. Defaults to \code{TRUE}.}

\item{title_size}{Font size for the plot title. Defaults to 20.}

\item{subtitle_size}{Font size for the plot subtitle. Defaults to 16.}

\item{axis_title_size}{Font size for the plot x- and y-axis titles. Defaults to 18.}

\item{axis_size}{Font size for the plot axis tick labels. Defaults to 18.}
}
\description{
This function allows you to add the Exploristics theme to a ggplot2 figure.
}
\details{
This function is added to your ggplot to set elements of the \code{theme()} including background,
axis, legend and titles to the style used at Exploristics.
}
\examples{
library(ggplot2)

## generate a plot and add the Exploristics theme
ggplot(data = mtcars, aes(x = hp, y = mpg)) +
geom_point() +
exploristics_theme()
}
\seealso{
\code{\link[ggplot2]{theme}}
}
