% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileVersioning.R
\name{FileVersioning}
\alias{FileVersioning}
\title{Creates a unique file pathway using simple versioning system}
\format{
Date and time format argument passed to \code{\link[base]{format}}
}
\usage{
FileVersioning(file_path, suffix = "integer", format = NULL, ...)
}
\arguments{
\item{file_path}{either A character string naming a file}

\item{suffix}{The suffix to append to the basename of file. Can be one of
integer, date, datetime or unix timestamp. See Details for more information.}
}
\value{
The supplied dataframe with standarised terminology
}
\description{
Wrapper function for file path to check if it already exists and
if so appends a suffix to the file name for versioning purposes.
}
\seealso{
\code{\link[base]{format}}
}
\author{
Gareth Burns
}
