% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_with_logo.R
\name{save_with_logo}
\alias{save_with_logo}
\title{Save a ggplot2 and add a footer with the Exploristics logo.}
\usage{
save_with_logo(
  plot = NULL,
  filename = NULL,
  text = NULL,
  line = FALSE,
  logo = NULL,
  width = 8,
  height = 8,
  dpi = 300,
  suffix = NULL
)
}
\arguments{
\item{plot}{A ggplot or the path to a previously saved image file.}

\item{filename}{The file to save the image to. Don't include file extension as it will be saved as a \code{.png}.}

\item{text}{Optional text to add to the left of the footer. Single line or across 2 lines if you add \verb{\\n} between the lines of text.}

\item{line}{Add a line between the plot and the footer. Defaults to \code{FALSE}.}

\item{logo}{Path to an image file to use in the footer. The path must be URL, filename or raw vector. Defaults uses the Exploristics logo supplied with the package.}

\item{width}{Width of the plot to save, in inches. Only used if \code{plot} is a ggplot. Defaults to \code{8}.}

\item{height}{Height of the plot to save, in inches. Only used if \code{plot} is a ggplot. Defaults to \code{8}.}

\item{dpi}{Dots per inch to use when saving the plot. Only used if \code{plot} is a ggplot. Defaults to \code{300}.}

\item{suffix}{Suffix to add to the original plot filename. Only used if \code{plot} is an image file. Don't need to include file extension. Defaults to "_with_footer".}
}
\description{
This function saves a ggplot with a footer with the Exploristics logo and an optional text caption added below the plot.
}
\details{
This function saves a ggplot with a footer with the Exploristics logo at the bottom-right corner and an optional text caption at the bottom-left corner below the plot.
There is also an option to add a line between the plot and the footer.
The plot will be saved with the footer added is saved to a the provided filename as a \code{.png} file.
An image file can also be used with this function. It will add the footer to an image then save it to a .png file with the same filename as the original plot image but with the suffix "with_footer" added to it.
}
\examples{
library(ggplot2)
library(magick)
library(magrittr)
library(scales)
library(stringr)

## generate a plot with the Exploristics theme
cars_plot <- ggplot(data = mtcars, aes(x = hp, y = mpg, colour = mpg)) +
geom_point(size=2) +
labs(title="Example plot") +
exploristics_theme()

## add colour scheme, wrap text labels and save with the Exploristics logo
cars_plot \%>\%
exploristics_colour(colour_pal="Expl_External") \%>\%
text_wrapper() \%>\%
save_with_logo(filename="example_cars_plot", text= "Source:Data source\nProduced by: Name",
width = 8, height = 8, dpi = 300)


## Alternatively, add the Exploristics logo to a previously saved plot ##

## generate a plot with the Exploristics theme
cars_plot <- ggplot(data = mtcars, aes(x = hp, y = mpg, colour = mpg)) +
geom_point(size=2) +
labs(title="Example plot") +
exploristics_theme()

## add colour scheme and wrap text labels
cars_plot \%>\%
exploristics_colour(colour_pal="Expl_External") \%>\%
text_wrapper()

## save the plot
ggsave(filename = "example_cars_plot.png", width = 8, height = 8, dpi = 300)

## add the footer to the saved image
save_with_logo("example_cars_plot.png", text= "Source:Data source\nProduced by: Name")
}
\seealso{
\code{\link[ggplot2]{ggsave}}

\code{\link[magick]{image_read}}

\code{\link[magick]{image_trim}}

\code{\link[magick]{image_scale}}

\code{\link[magick]{image_background}}

\code{\link[magick]{image_extent}}

\code{\link[magick]{image_border}}

\code{\link[magick]{image_annotate}}

\code{\link[magick]{image_info}}

\code{\link[magick]{image_blank}}

\code{\link[magick]{image_append}}

\code{\link[magick]{image_write}}
}
