% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_exploristics.R
\name{scale_fill_exploristics}
\alias{scale_fill_exploristics}
\title{Exploristics Gradient fill scales}
\usage{
scale_fill_exploristics()
}
\description{
Exploristics Gradient fill scales
}
\seealso{
\code{\link[ggplot2]{scale_color_gradient}}
}
