% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploristics_theme.R
\name{theme_exploristics}
\alias{theme_exploristics}
\title{Add Exploristics theme to ggplot2 figure.}
\usage{
exploristics_theme(
  base_size = 12,
  base_family = "",
  base_line_size = base_size/11,
  base_rect_size = base_size/11
)
}
\arguments{
\item{base_size}{base font size, given in pts.}

\item{base_family}{base font family}

\item{base_line_size}{base size for line elements}

\item{base_rect_size}{base size for rect elements}
}
\description{
A \code{\link[ExploristicsTheme]{exploristics_theme}} function but
starting with \code{theme} naming convention.
}
\details{
The \code{theme_} prefix is required for use in KerusCloud VPV and
having both \code{\link[ExploristicsTheme]{theme_exploristics}} and
\code{\link[ExploristicsTheme]{exploristics_theme}} maintains backwards compatibility
}
\seealso{
\code{\link[ggplot2]{theme}}
}
\author{
Gareth Burns
}
