% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Expl_Blue}
\alias{Expl_Blue}
\title{Colour Pallette Using Exploristics Blues}
\format{
An object of class \code{character} of length 6.
}
\usage{
Expl_Blue
}
\description{
Colour Pallette Using Exploristics Blues
}
\details{
Contains the following hex colours
\itemize{
\item{#002659}
\item{#293D94}
\item{#136BBA}
\item{#16A4D4}
\item{#9DD6E4}
\item{#D6EFF5}
}
}
\keyword{datasets}
