% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Expl_Brand}
\alias{Expl_Brand}
\title{Colour Pallette Using Exploristics Brand Colours}
\format{
An object of class \code{character} of length 12.
}
\usage{
Expl_Brand
}
\description{
Colour Pallette Using Exploristics Brand Colours
}
\details{
Contains the following hex colours
\itemize{
\item{#121a31"}
\item{#2d3750"}
\item{#cfdbef"}
\item{#6ddabf"}
\item{#4c9d88"}
\item{#ae6867"}
\item{#d8807f"}
\item{#ca9e8e"}
\item{#f6c0ad"}
\item{#b1ca65"}
\item{#d7f67c"}
\item{#576fb8"}
}
}
\keyword{datasets}
