% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Expl_External}
\alias{Expl_External}
\title{Colour Pallette Using A Colour Palette for External Clients}
\format{
An object of class \code{character} of length 6.
}
\usage{
Expl_External
}
\description{
Colour Pallette Using A Colour Palette for External Clients
}
\details{
Contains the following hex colours
\itemize{
\item{#7119A2}
\item{#28CAAB}
\item{#3C56E1}
\item{#B04AFF}
\item{#97EDDB}
\item{#92C8FF}
}
}
\keyword{datasets}
