% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Expl_HighCont}
\alias{Expl_HighCont}
\title{Colour Pallette Using High Contrast Colours}
\format{
An object of class \code{character} of length 6.
}
\usage{
Expl_HighCont
}
\description{
Colour Pallette Using High Contrast Colours
}
\details{
Contains the following hex colours
\itemize{
\item{#2D2669}
\item{#FF4A1B}
\item{#18CDFF}
\item{#E8E126}
\item{#BF00B5}
\item{#5EB17D}
}
}
\keyword{datasets}
