% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Expl_RGB}
\alias{Expl_RGB}
\title{Colour Pallette of 6 colours including blues, reds, green and yellow}
\format{
An object of class \code{character} of length 6.
}
\usage{
Expl_RGB
}
\description{
Colour Pallette of 6 colours including blues, reds, green and yellow
}
\details{
Contains the following hex colours
\itemize{
\item{#6b89e3"}
\item{#d8807f"}
\item{#d7f67c"}
\item{#6ddabf"}
\item{#9eb7de"}
\item{#f6c0ad"}
}
}
\keyword{datasets}
