% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploristics_colour.R
\name{exploristics_colour}
\alias{exploristics_colour}
\title{Add Exploristics colour to a ggplot2 figure.}
\usage{
exploristics_colour(
  plot,
  colour_pal = c("Expl_Blue", "Expl_External", "Expl_HighCont"),
  rev_gradient = F
)
}
\arguments{
\item{plot}{A ggplot2 figure (S3: gg, ggplot).}

\item{colour_pal}{The colour palette to use. Can be \code{Expl_Blue}, \code{Expl_External} or \code{Expl_HighCont}. Defaults to \code{Expl_Blue}}

\item{rev_gradient}{Should the reverse gradient of the colour palette be used (only for continuous variables). Defaults to \code{FALSE}.}
}
\description{
This function returns a ggplot with an Exploristics colour scheme added to it.
There are 3 colour palettes to choose from.They are suitable for those with most common types of colour blindness.
}
\details{
This function sets the colour of the ggplot to a palette matching an
Exploristics colour scheme. A gradient of the palette will be applied for continuous variables.
\code{Expl_Blue} is a palette of 6 blues ranging from a dark navy to a light green-blue.
\code{Expl_External} is a palette of 6 colours with purples, greens and blues.
\code{Expl_HighCont} is a high-contrast palette of 6 colours with a variety of colours.
}
\examples{
library(ggplot2)

## generate a plot with the Exploristics theme
cars_plot <- ggplot(data = mtcars, aes(x = hp, y = mpg, colour = mpg)) +
geom_point(size=2) +
exploristics_theme()

## adding the colour scheme can be written as one line
exploristics_colour(cars_plot, colour_pal="Expl_External")


library(magrittr)

## or using a pipe
cars_plot \%>\%
exploristics_colour(colour_pal="Expl_External")
}
\seealso{
\code{\link[ggplot2]{scale_color_manual}}

\code{\link[ggplot2]{scale_color_gradientn}}
}
