% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploristics_fill.R
\name{exploristics_fill}
\alias{exploristics_fill}
\title{Add Exploristics fill to ggplot chart}
\usage{
exploristics_fill()
}
\description{
This function allows you to add the Exploristics fill to your ggplotgraphics.
This function sets the fill of the graphs to a gradient matching the exploristics colour scheme
}
\keyword{exploristics_fill}
