% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_colour_exploristics.R
\name{scale_colour_exploristics}
\alias{scale_colour_exploristics}
\title{Exploristics Gradient colour scales}
\usage{
scale_colour_exploristics()
}
\description{
Exploristics Gradient colour scales
}
\seealso{
\code{\link[ggplot2]{scale_color_gradient}}
}
