% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_exploristics.R
\name{scale_fill_exploristics}
\alias{scale_fill_exploristics}
\title{Exploristics Gradient fill scales}
\usage{
scale_fill_exploristics(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[ggplot2]{scale_color_gradient}}}
}
\description{
A two colour gradient for continuous variables to apply to fill aesthetic
}
\examples{
library(ggplot2)
ggplot(data = mtcars, aes(x = wt, y = mpg, size = hp, fill = mpg)) +
geom_point(shape = 21) +
scale_colour_exploristics()
}
\seealso{
\code{\link[ggplot2]{scale_color_gradient}}
}
\author{
Gareth Burns
}
