% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_exploristics.R
\name{theme_exploristics}
\alias{theme_exploristics}
\title{Add Exploristics theme to ggplot2 figure.}
\usage{
theme_exploristics(
  base_size = 12,
  base_family = "",
  base_line_size = base_size/11,
  base_rect_size = base_size/11
)
}
\arguments{
\item{base_size}{base font size, given in pts.}

\item{base_family}{base font family}

\item{base_line_size}{base size for line elements}

\item{base_rect_size}{base size for rect elements}
}
\description{
This function allows you to add the Exploristics theme to a ggplot2 figure.
}
\details{
This function is added to your ggplot to set elements of the \code{`theme()`}
including background, axis, legend and titles to the style used at Exploristics.
}
\examples{
library(ggplot2)

## generate a plot and add the Exploristics theme
ggplot(data = mtcars, aes(x = hp, y = mpg)) +
geom_point() +
theme_exploristics()
}
\seealso{
\code{\link[ggplot2]{theme}}
}
\author{
Amy McCorry
}
